Move the ResourceFinder into the Config folder
