+++ src/Config/ResourceFinder.php
@@ -8,7 +8,7 @@
  * @license LGPL-3.0+
  */
 
-namespace Contao\CoreBundle\Finder;
+namespace Contao\CoreBundle\Config;
 
 use Symfony\Component\Finder\Finder;
